split currency to its own module
